+++ spec/array_spec.lua
@@ -154,7 +154,31 @@ context('Array functions specs', function()
     end)  
     
   end)
-
+	
+  context('findIndex', function()
+  
+    test('returns the first index at which a predicate passes a truth test', function()
+      assert_equal(_.findIndex({1,2,3,4,5},function(__,v) return v%2==0 end),2)
+    end)
+    
+    test('returns nil when nothing was found', function()
+      assert_nil(_.findIndex({1,2,3,4,5},function(_,v) return v>5 end))
+    end)  
+    
+  end)	
+	
+  context('findLastIndex', function()
+  
+    test('returns the last index at which a predicate passes a truth test', function()
+      assert_equal(_.findLastIndex({1,2,3,4,5},function(_,v) return v%2==0 end),4)
+    end)
+    
+    test('returns nil when nothing was found', function()
+      assert_nil(_.findLastIndex({1,2,3,4,5},function(_,v) return v>5 end))
+    end)  
+    
+  end)
+	
   context('addTop', function()
   
     test('adds values at the top of an array', function()